Fixed issues with Conversation selection and getting Messages back from the database
