+++ src/codeu/chat/client/ClientConversation.java
@@ -20,7 +20,7 @@ import java.util.Map;
 
 import codeu.chat.common.Conversation;
 import codeu.chat.common.ConversationSummary;
-import codeu.chat.common.EncryptionKey;
+import codeu.chat.util.EncryptionKey;
 import codeu.chat.util.Logger;
 import codeu.chat.util.Method;
 import codeu.chat.util.Uuid;
