+++ src/codeu/chat/client/ClientMessage.java
@@ -22,11 +22,8 @@ import java.util.Map;
 
 import codeu.chat.common.Conversation;
 import codeu.chat.common.ConversationSummary;
-import codeu.chat.common.EncryptionKey;
+import codeu.chat.util.*;
 import codeu.chat.common.Message;
-import codeu.chat.util.Logger;
-import codeu.chat.util.Method;
-import codeu.chat.util.Uuid;
 
 public final class ClientMessage {
