+++ src/codeu/chat/client/maingui/MessagePanel.java
@@ -9,6 +9,7 @@ import java.math.BigInteger;
 
 import codeu.chat.client.ClientContext;
 import codeu.chat.common.*;
+import codeu.chat.util.RSA;
 
 // NOTE: JPanel is serializable, but there is no need to serialize MessagePanel
 // without the @SuppressWarnings, the compiler will complain of no override for serialVersionUID
@@ -142,8 +143,8 @@ public final class MessagePanel extends JPanel{
           JOptionPane.showMessageDialog(MessagePanel.this, "You must select a conversation.");
         } else {
           final String messageText = sendMessageTextArea.getText();
+
           if (messageText != null && messageText.length() > 0) {
-            // Implement sending message with file here:
             clientContext.message.addMessage(
                 clientContext.user.getCurrent().id,
                 clientContext.conversation.getCurrentId(),
@@ -195,23 +196,19 @@ public final class MessagePanel extends JPanel{
   // Populate ListModel
   private void getAllMessages(ConversationSummary conversation) {
     messageListModel.clear();
-
     for (final Message m : clientContext.message.getConversationContents(conversation)) {
       // Display author name if available.  Otherwise display the author UUID.
       final String authorName = clientContext.user.getName(m.author);
       try {
-        BigInteger encryptedContent = RSA.keyToBigInteger(m.content);
+        BigInteger encryptedContent = RSA.valueToBigInteger(m.content);
         Conversation currentConversation = clientContext.conversation.getCurrentConversation();
-        System.out.println("getAllMessages " + currentConversation);
         m.content = RSA.messageToString(RSA.decrypt(encryptedContent, currentConversation.SecretKey()));
         final String displayString = String.format("@%s: %s",
                 ((authorName == null) ? m.author : authorName), m.content);
-
         messageListModel.addElement(displayString);
-      }catch (NumberFormatException ex){
+      }catch (NumberFormatException nfe){
         final String displayString = String.format("@%s: %s",
                 ((authorName == null) ? m.author : authorName), m.content);
-
         messageListModel.addElement(displayString);
       }
     }
