+++ src/codeu/chat/server/Controller.java
@@ -14,25 +14,13 @@
 
 package codeu.chat.server;
 
-import java.util.Collection;
-import java.util.List;
-import java.util.Arrays;
-import java.util.HashMap;
-
 import codeu.chat.common.*;
+import codeu.chat.util.EncryptionKey;
 import codeu.chat.util.Time;
 import codeu.chat.util.Uuid;
-import jdk.nashorn.internal.ir.ReturnNode;
 import codeu.chat.util.Logger;
 
-import java.io.BufferedReader;
-import java.io.FileReader;
 import java.io.File;
-import java.io.IOException;
-import java.io.FileNotFoundException;
-
-import java.nio.file.Files;
-import java.nio.file.Paths;
 
 public final class Controller implements RawController, BasicController {
 
@@ -95,21 +83,19 @@ public final class Controller implements RawController, BasicController {
 
         // The conversation has no messages in it, that's why the last message is NULL (the first
         // message should be NULL too. Since there is no last message, then it is not possible
-        // to update the last message's "next" value.
+        // to update the last message's "next" value. But we must add this message as the new
+        // last and first messages.
+
+        foundConversation.firstMessage = message.id;
 
       } else {
         final Message lastMessage = model.messageById().first(foundConversation.lastMessage);
         lastMessage.next = message.id;
+        message.previous = lastMessage.id;
+
+        persistanceController.addMessage(lastMessage, foundConversation);
+
       }
-
-      // If the first message points to NULL it means that the conversation was empty and that
-      // the first message should be set to the new message. Otherwise the message should
-      // not change.
-
-      foundConversation.firstMessage =
-          Uuid.equals(foundConversation.firstMessage, Uuid.NULL) ?
-          message.id :
-          foundConversation.firstMessage;
 
       // Update the conversation to point to the new last message as it has changed.
