+++ src/codeu/chat/util/Serializers.java
@@ -14,9 +14,6 @@
 
 package codeu.chat.util;
 
-import codeu.chat.common.EncryptionKey;
-import codeu.chat.common.RSA;
-
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
